Move KButton visual test to visual.spec.js
