+++ lib/buttons-and-links/__tests__/KButton.spec.js
@@ -2,7 +2,6 @@ import { render, screen } from '@testing-library/vue';
 import userEvent from '@testing-library/user-event';
 import VueRouter from 'vue-router';
 import KButton from '../KButton.vue';
-import { renderComponentForVisualTest, takeSnapshot } from '../../../jest.conf/visual.testUtils';
 
 const renderComponent = (props = {}, slots = {}) =>
   render(KButton, {
@@ -60,14 +59,6 @@ describe('KButton', () => {
       const clickEvent = emitted();
       expect(clickEvent).toHaveProperty('click');
       expect(clickEvent.click).toHaveLength(1);
-    });
-  });
-
-  describe.visual('KButton visual tests', () => {
-    const snapshotOptions = { widths: [800], minHeight: 512 };
-    it('renders correctly with different appearances', async () => {
-      await renderComponentForVisualTest('KButtonVisualTest');
-      await takeSnapshot('KButton visual tests', snapshotOptions);
     });
   });
 });
